requirements.txt
numpy==1.21.2
pandas==1.2.0
plotly==5.1.0
dash==2.0.0
psycopg2-binary==2.8.5
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
dash-bootstrap-components>=0.10.1
dash-leaflet==0.1.23
click==7.1.2
Flask==1.1.2
Flask-Compress==1.10.1
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
dash_daq==0.5.0
matplotlib==3.3.4
requests==2.28.0
xlsxwriter == 1.3.8
geopy == 2.0.0
